+++ ui/src/app/job-details/tabs/tabs.component.ts
@@ -61,6 +61,20 @@ export class JobTabsComponent implements OnInit, OnChanges {
     return JobStatusIcon[status];
   }
 
+  hasCallCachedTask(): boolean {
+    if (this.tasks && this.tasks.find(t => t.callCached === true)) {
+      return true;
+    }
+    return false;
+  }
+
+  hasScatteredTask(): boolean {
+    if (this.tasks && this.tasks.find(t => t.shards !== null)) {
+      return true;
+    }
+    return false;
+  }
+
   hasFailures(): boolean {
     return this.job.failures && (this.job.failures.length !== 0);
   }